refactor: mv API middleware/types to /lib/api
