+++ src/app/api/accounts/route.ts
@@ -1,10 +1,9 @@
 import { NextResponse } from "next/server";
 import { eq } from "drizzle-orm";
 
+import { withAuth } from "@/lib/api/middleware";
 import { db } from "@/lib/db";
 import { monzoAccounts } from "@/lib/db/schema/monzo-schema";
-
-import { withAuth } from "../middleware";
 
 export const GET = withAuth<typeof monzoAccounts.$inferSelect>(
   async ({ userId }) => {
