+++ src/app/api/categories/[id]/merchants/route.ts
@@ -1,9 +1,9 @@
 import { NextResponse } from "next/server";
 import { and, eq } from "drizzle-orm";
 
+import { withAuth } from "@/lib/api/middleware";
 import { db } from "@/lib/db";
 import { monzoMerchants } from "@/lib/db/schema/monzo-schema";
-import { withAuth } from "@/app/api/middleware";
 
 // NOTE: this is a POST request even though we are fetching because of the request body
 export const POST = withAuth<
